requirements.txt
streamlit
openai-whisper
fairseq2
seamless_communication @ git+https://github.com/facebookresearch/seamless_communication@main